# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentCordon API
  version: 1.6.1
  description: |
    AgentCordon is a self-hostable Agentic Identity Provider (IdP) that provides
    secure, auditable, policy-controlled token brokerage for autonomous agents.
    Agents do not directly hold long-lived credentials; every issuance and use
    is observable via the audit log.
  contact:
    name: AgentCordon
  license:
    name: MIT
servers:
  - url: http://localhost:3140
    description: Local development server
...
security:
  - SessionCookie: []
  - BearerJWT: []
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: agtcrdn_session
      description: |
        Session cookie set after user login. Every unsafe method authenticated
        this way is also CSRF-protected by a double-submit rule: the request
        must carry an `X-CSRF-Token` header whose value matches the readable
        `agtcrdn_csrf` cookie, or it is refused with
        `403 csrf_validation_failed`.
    BearerJWT:
      type: http
      scheme: bearer
      description: |
        An **opaque** OAuth 2.0 access token issued to a workspace by
        `POST /api/v1/oauth/token`. Despite the scheme name this is not a JWT
        and carries no claims a client can read: the server stores only its
        SHA-256 hash and resolves the workspace by token id. The ES256 JWT
        issuer, its JWKS endpoint, and the permissions-token endpoint were
        removed.
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
        data:
          description: Response payload (present when status is ok)
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    CredentialPermissionsResponse:
      type: object
      properties:
        credential_id:
          type: string
          format: uuid
        owner_agent:
          type: string
          format: uuid
          nullable: true
        owner_user:
          type: string
          format: uuid
          nullable: true
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionEntry'
    PermissionEntry:
      type: object
      properties:
        agent_id:
          type: string
          format: uuid
        permission:
          type: string
          enum: [read, write, delete, delegated_use]
        granted_by:
          type: string
          format: uuid
          nullable: true
        granted_by_user:
          type: string
          format: uuid
          nullable: true
        granted_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/credentials/{id}/permissions:
    get:
      tags: [Permissions]
      summary: Get credential permissions
      security:
        - BearerJWT: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Credential permissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CredentialPermissionsResponse'
        '404':
          description: Credential not found
        '403':
          description: Access denied by policy

    post:
      tags: [Permissions]
      summary: Grant permission(s) to an agent
      description: Supports single (permission field) or batch (permissions array) grants.
      security:
        - BearerJWT: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id]
              properties:
                agent_id:
                  type: string
                  format: uuid
                permission:
                  type: string
                  enum: [read, write, delete, delegated_use]
                permissions:
                  type: array
                  items:
                    type: string
                    enum: [read, write, delete, delegated_use]
      responses:
        '200':
          description: Permission(s) granted
        '400':
          description: Invalid permission or missing required fields
        '404':
          description: Credential or agent not found

    put:
      tags: [Permissions]
      summary: Set permissions (replace all)
      description: Replaces all permissions for a credential with the provided list.
      security:
        - BearerJWT: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissions]
              properties:
                permissions:
                  type: array
                  items:
                    type: object
                    required: [agent_id, permission]
                    properties:
                      agent_id:
                        type: string
                        format: uuid
                      permission:
                        type: string
                        enum: [read, write, delete, delegated_use]
      responses:
        '200':
          description: Permissions set
  /api/v1/credentials/{id}/permissions/{agent_id}/{permission}:
    delete:
      tags: [Permissions]
      summary: Revoke a specific permission
      security:
        - BearerJWT: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: permission
          in: path
          required: true
          schema:
            type: string
            enum: [read, write, delete, delegated_use]
      responses:
        '200':
          description: Permission revoked

  # ============================================================================
  # Grants
  # ============================================================================
...
  /api/v1/mcp-servers/{id}/permissions/{agent_id}/{permission}:
    delete:
      tags: [MCP]
      summary: Revoke MCP server permission
      description: |
        Revokes an agent's permission on this MCP server.
        Requires manage_mcp_servers policy.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: permission
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission revoked
        '404':
          description: MCP server or permission not found

  # ── MCP OAuth Apps ─────────────────────────────────────────────────────────
...